move retail react app jest setup
