+++ packages/template-retail-react-app/jest-setup.js
@@ -6,10 +6,65 @@
  */
 const path = require('path')
 const mockConfig = require(path.join(__dirname, 'config/mocks/default.js'))
+require('raf/polyfill') // fix requestAnimationFrame issue with polyfill
+require('@testing-library/jest-dom/extend-expect')
+const {Crypto} = require('@peculiar/webcrypto')
 
 // Mock the application configuration to be used in all tests.
 jest.mock('pwa-kit-runtime/utils/ssr-config', () => {
     return {
-        getConfig: () => mockConfig
+        getConfig: () => mockConfig,
     }
+})
+
+// TextEncoder is a web API, need to import it
+// from nodejs util in testing environment.
+// This is used in commerce-api/pkce.js
+global.TextEncoder = require('util').TextEncoder
+
+// This file consists of global mocks for jsdom.
+class LocalStorageMock {
+    constructor() {
+        this.store = {}
+    }
+    clear() {
+        this.store = {}
+    }
+    getItem(key) {
+        return this.store[key] || null
+    }
+    setItem(key, value) {
+        this.store[key] = value?.toString()
+    }
+    removeItem(key) {
+        delete this.store[key]
+    }
+}
+
+const localStorageMock = new LocalStorageMock()
+
+Object.defineProperty(window, 'crypto', {
+    value: new Crypto(),
+})
+
+Object.defineProperty(window, 'localStorage', {
+    value: localStorageMock,
+})
+
+Object.defineProperty(window, 'scrollTo', {
+    value: () => null,
+})
+
+Object.defineProperty(window, 'matchMedia', {
+    writable: true,
+    value: jest.fn().mockImplementation((query) => ({
+        matches: false,
+        media: query,
+        onchange: null,
+        addListener: jest.fn(), // deprecated
+        removeListener: jest.fn(), // deprecated
+        addEventListener: jest.fn(),
+        removeEventListener: jest.fn(),
+        dispatchEvent: jest.fn(),
+    })),
 })
